hidl: camera: camera3: Change usage type to uint64_t
* gralloc BufferUsage is already uint64_t, this seems to be samsung's real change here

+++ hidl/camera/libhardware_headers/include/hardware/camera3.h
@@ -1657,7 +1657,7 @@ typedef struct camera3_stream {
      *   supported due to imcompatible buffer format, dataSpace, or other hardware
      *   limitations.
      */
-    uint32_t usage;
+    uint64_t usage;
 
     /**
      * The maximum number of buffers the HAL device may need to have dequeued at
